+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Target Scheduler
 
-## 5.0.3.0 (beta) - 2025-02-XX
-* Fixed display and threading problems with Scheduler Preview
+## 5.1.0.0 (beta) - 2025-02-XX
+* Planner will now attempt to stick with a selected target for the minimum time.
+* Fixed display and threading problems with Scheduler Preview.
 
 ## 5.0.2.0 (beta) - 2025-02-12
 * Fixed Telescopius CSV format change.